chore: Rename `PlaybackSevice.ts`
- Mainly because we have a `Playback.ts` service for the media module
