+++ mobile/index.ts
@@ -4,7 +4,7 @@ import "intl-pluralrules";
 import "expo-router/entry";
 import TrackPlayer from "react-native-track-player";
 
-import { PlaybackService, setupPlayer } from "./src/services/PlaybackService";
+import { PlaybackService, setupPlayer } from "./src/services/RNTPService";
 
 /*
   Create custom entry point to potentially fix issues with
